Git fixes & cleanup
