--- Virtual-Game-Shelf-master/README.md
@@ -1,25 +1,0 @@
-# Virtual Game Shelf - A Game Collection Organizer
-Virtual Game Shelf is a desktop Java application to organize your 
-game library in a single convenient location. It strives to implement 
-the best features from services like [Backloggery](http://www.backloggery.com/), 
-without the need for an internet connection.
-
-Code contributions should follow [Google's Java Style Guide](https://google.github.io/styleguide/javaguide.html)
-
-## Compiling with BlueJ
-
-In BlueJ, Select **`Project -> Open Non BlueJ...`**. Select the project 
-folder in the newly opened window and click **`Open in BlueJ`**. 
-(By default, the folder should be named **Virtual-Game-Shelf**.) 
-You should now be able to open and compile the project in BlueJ.
-
-Note: A Package Line Mismatch warning may appear if project 
-files have been recently moved or renamed. If this happens, 
-click "continue" to have BlueJ automatically correct the issue.
-
-## Automatic testing with Travis CI
-
-All pull requests are [automatically test compiled](https://travis-ci.com/Stevoisiak/Virtual-Game-Shelf/) 
-in Travis CI using an Ant `build.xml` file generated by Eclipse.
-
-Travis CI specific settings are defined in `.travis.yml`.
