--- Virtual-Game-Shelf-master/src/VirtualGameShelf.ctxt
@@ -1,22 +1,0 @@
-#BlueJ class context
-comment0.params=args
-comment0.target=void\ main(java.lang.String[])
-comment1.params=mainStage
-comment1.target=void\ start(javafx.stage.Stage)
-comment2.params=
-comment2.target=javafx.scene.control.MenuButton\ generateMenuButton()
-comment3.params=observable\ oldValue\ newValue
-comment3.target=void\ changed(javafx.beans.value.ObservableValue,\ java.lang.Boolean,\ java.lang.Boolean)
-comment4.params=
-comment4.target=javafx.scene.control.MenuBar\ generateMenuBar()
-comment5.params=
-comment5.target=void\ manualAdd()
-comment6.params=ov\ oldValue\ newValue
-comment6.target=void\ changed(javafx.beans.value.ObservableValue,\ java.lang.String,\ java.lang.String)
-comment7.params=ov\ oldValue\ newValue
-comment7.target=void\ changed(javafx.beans.value.ObservableValue,\ java.lang.String,\ java.lang.String)
-comment8.params=observable\ oldValue\ newValue
-comment8.target=void\ changed(javafx.beans.value.ObservableValue,\ java.lang.String,\ java.lang.String)
-comment9.params=e
-comment9.target=void\ handle(javafx.event.ActionEvent)
-numComments=10
