--- Virtual-Game-Shelf-master/src/assets/stylesheet.css
@@ -1,15 +1,0 @@
-.menu-button
-{
-	-fx-border-color: transparent;
-    -fx-border-width: 0;
-    -fx-background-radius: 0;
-    -fx-background-color: transparent;
-}
-
-.menu-button > .arrow-button {
-    -fx-padding: 0;
-}
-
-.menu-button > .arrow-button > .arrow {
-    -fx-padding: 0;
-}
