--- Virtual-Game-Shelf-master/src/package.bluej
@@ -1,25 +1,0 @@
-#BlueJ package file
-objectbench.height=76
-objectbench.width=656
-package.editor.height=398
-package.editor.width=558
-package.editor.x=681
-package.editor.y=3
-package.numDependencies=0
-package.numTargets=1
-package.showExtends=true
-package.showUses=true
-project.charset=windows-1252
-target1.editor.height=694
-target1.editor.width=899
-target1.editor.x=471
-target1.editor.y=43
-target1.height=50
-target1.name=VirtualGameShelf
-target1.naviview.expanded=false
-target1.showInterface=false
-target1.type=ClassTarget
-target1.typeParameters=
-target1.width=120
-target1.x=160
-target1.y=10
