+++ src/assets/stylesheet.css
@@ -1,5 +1,5 @@
 .menu-button {
-	-fx-border-color: transparent;
+    -fx-border-color: transparent;
     -fx-border-width: 0;
     -fx-background-radius: 0;
     -fx-background-color: transparent;
